Move envvar package from cmd/frontend to internal.
