+++ internal/conf/auth.go
@@ -1,7 +1,7 @@
 package conf
 
 import (
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
+	"github.com/sourcegraph/sourcegraph/internal/envvar"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
