+++ internal/profiler/profiler.go
@@ -3,8 +3,8 @@ package profiler
 import (
 	"cloud.google.com/go/profiler"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
 	"github.com/sourcegraph/sourcegraph/internal/env"
+	"github.com/sourcegraph/sourcegraph/internal/envvar"
 	"github.com/sourcegraph/sourcegraph/internal/version"
 )
